+++ database/migrations/2021_06_15_093407_create_apartments_table.php
@@ -25,7 +25,8 @@ class CreateApartmentsTable extends Migration
             $table -> string('url_img');
             $table -> string('features');
 
-            $table -> integer('landlord_id')->index()->unsigned();
+            $table -> bigInteger('landlord_id')->index()->unsigned();
+
 
             $table->timestamps();
         });
